Test on Kirin950 cortex-A53 (1-thread, theoretical 6.0 GFLOPS)
| M, N, K | fp64 GFLOPS before this MR | fp64 GFLOPS after this MR |
| ---------- | --------------------------------------- | ------------------------------------- |
| 300,300,300 | 1.7 | 3.4 |
| 500,500,500 | 1.7 | 3.5 |

Test on Amlogic S905D3 (1-thread, theoretical 7.6 GFLOPS)
| M, N, K | fp64 GFLOPS before this MR | fp64 GFLOPS after this MR |
| ---------- | --------------------------------------- | ------------------------------------- |
| 300,300,300 | 2.1 | 4.3 |
| 500,500,500 | 2.2 | 4.6 |

Test program: 
[OpenBLAS-zgemm-test.tar.gz](https://github.com/xianyi/OpenBLAS/files/7596447/OpenBLAS-zgemm-test.tar.gz)
(test_kernel.cpp for correctness of kernel function (with all 4 conjugate cases), test_zgemm.cpp for correctness and performance of zgemm)